requirement.txt
requirement : 
discord.py
python-dotenv
this sucks actually ->argostranslate